requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-dateutil>=2.8.0
schedule>=1.2.0
urllib3>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
discord-webhook==1.3.0
APScheduler==3.10.4
lxml==4.9.3
pytz==2023.3
coverage==7.3.2
# Authentication dependencies
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.0 